[naga] Move `HandleSet` from `compact` into `arena`.
