+++ naga/src/compact/functions.rs
@@ -1,4 +1,4 @@
-use super::handle_set_map::HandleSet;
+use super::arena::HandleSet;
 use super::{FunctionMap, ModuleMap};
 
 pub struct FunctionTracer<'a> {
