+++ naga/src/compact/mod.rs
@@ -4,8 +4,9 @@ mod handle_set_map;
 mod statements;
 mod types;
 
+use crate::arena::HandleSet;
 use crate::{arena, compact::functions::FunctionTracer};
-use handle_set_map::{HandleMap, HandleSet};
+use handle_set_map::HandleMap;
 
 /// Remove unused types, expressions, and constants from `module`.
 ///
